+++ cmd/helpers.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"path/filepath"
 	"strings"
 
 	"github.com/freshly/tuber/graph/model"
@@ -16,7 +15,6 @@ import (
 	"github.com/freshly/tuber/pkg/report"
 
 	"github.com/getsentry/sentry-go"
-	"github.com/goccy/go-yaml"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -157,94 +155,6 @@ func credentials() ([]byte, error) {
 	}
 
 	return creds, nil
-}
-
-type tuberConfig struct {
-	Clusters []Cluster
-}
-
-// Cluster is a cluster
-type Cluster struct {
-	Name      string `yaml:"name"`
-	Shorthand string `yaml:"shorthand"`
-	URL       string `yaml:"url"`
-}
-
-func (c tuberConfig) CurrentClusterConfig() Cluster {
-	name, err := k8s.CurrentCluster()
-	if err != nil {
-		return Cluster{}
-	}
-
-	return c.FindByName(name)
-}
-
-func (c tuberConfig) FindByShortName(name string) Cluster {
-	for _, cl := range c.Clusters {
-		if cl.Shorthand == name {
-			return cl
-		}
-	}
-
-	return Cluster{}
-}
-
-func (c tuberConfig) FindByName(name string) Cluster {
-	for _, cl := range c.Clusters {
-		if cl.Name == name {
-			return cl
-		}
-	}
-
-	return Cluster{}
-}
-
-func mustGetTuberConfig() *tuberConfig {
-	config, err := getTuberConfig()
-
-	if err != nil {
-		panic(err)
-	}
-
-	return config
-}
-
-func getTuberConfig() (*tuberConfig, error) {
-	path, err := tuberConfigPath()
-	if err != nil {
-		return nil, err
-	}
-
-	raw, err := ioutil.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-
-	var t tuberConfig
-	err = yaml.Unmarshal(raw, &t)
-	if err != nil {
-		return nil, err
-	}
-
-	return &t, nil
-}
-
-func tuberConfigPath() (string, error) {
-	dir, err := tuberConfigDir()
-	if err != nil {
-		return "", err
-	}
-
-	return filepath.Join(dir, "config.yaml"), nil
-}
-
-func tuberConfigDir() (string, error) {
-	basePath, err := os.UserConfigDir()
-	if err != nil {
-		return "", err
-	}
-
-	return filepath.Join(basePath, "tuber"), nil
 }
 
 func checkAuth() error {
