+++ cmd/resume.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/freshly/tuber/graph"
 	"github.com/freshly/tuber/graph/model"
+	"github.com/freshly/tuber/pkg/config"
 	"github.com/spf13/cobra"
 )
 
@@ -16,7 +17,7 @@ var resumeCmd = &cobra.Command{
 	PreRunE:      promptCurrentContext,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		appName := args[0]
-		graphql := graph.NewClient(mustGetTuberConfig().CurrentClusterConfig().URL)
+		graphql := graph.NewClient(config.MustLoad().CurrentClusterConfig().URL)
 
 		b := false
 		input := &model.AppInput{
